more changes to zip job

+++ lib/zip_recording.rb
@@ -13,8 +13,8 @@ class ZipRecording
   end
 
   def self.perform(label, type, files)
-    if !File.exist?("public/zips/#{label}.#{type}.zip")
-      t = File.open("public/zips/#{label}.#{type}.zip", "w")
+    if !File.exist?("public/zips/#{label}.#{type}.zip") && !File.exist?("public/zips/#{label}.#{type}.zip.tmp")
+      t = File.open("public/zips/#{label}.#{type}.zip.tmp", "w")
       Zip::ZipOutputStream.open(t.path) do |zos|
         files.each do |filename|
           file = File.open(filename)
@@ -23,6 +23,7 @@ class ZipRecording
         end
       end
       t.close
+      FileUtils.mv("public/zips/#{label}.#{type}.zip.tmp", "public/zips/#{label}.#{type}.zip")
     end
   end
 end
